rename firestore 'debug' to 'emulator, stream logs
