+++ apps/rates-service/src/internals/rates/db.test.ts
@@ -1,5 +1,5 @@
 import { init, Timestamp } from '@thecointech/firestore';
-import { filterByEmulator } from '@thecointech/firestore/debug';
+import { filterByEmulator } from '@thecointech/firestore/emulator';
 import { setRate, getCoinRate, toDbType } from './db';
 import { CoinRate } from "./types";
 import { describe } from '@thecointech/jestutils';
