+++ libs/broker-db/src/converter.test.ts
@@ -1,5 +1,5 @@
 import { getFirestore } from '@thecointech/firestore';
-import { init, filterByEmulator, Timestamp } from '@thecointech/firestore/debug';
+import { init, filterByEmulator, Timestamp } from '@thecointech/firestore/emulator';
 import { describe } from '@thecointech/jestutils';
 import { getUserData, setUserVerified } from './user';
 import { DateTime } from 'luxon';
