+++ libs/firestore/src/emulator.test.ts
@@ -1,8 +1,8 @@
-import { init, filterByEmulator } from './debug';
+import { init, filterByEmulator } from './emulator';
 import { describe } from '@thecointech/jestutils';
 import { getFirestore, Timestamp } from '.';
 
-
+console.log(process.env);
 //
 // Test the connection to the firestore emulator.
 //
@@ -34,4 +34,4 @@ describe('Our testing correctly connects to Firestore', () => {
     const ts = Timestamp.now();
     expect(ts).toBeTruthy();
   });
-}, filterByEmulator);
+}, filterByEmulator());
